refactor(view, controller): 수정 로또를 살 가격 입력 받기 기능
클래스의 흐름이 원활하도록 리팩토링을 진행하였습니다.
