+++ src/main/java/lotto/Application.java
@@ -2,11 +2,8 @@ package lotto;
 
 import lotto.controller.LottoController;
 import lotto.domain.Money;
-import lotto.domain.Store;
+import lotto.service.Store;
 import lotto.service.StatisticProducer;
-import lotto.validation.money.MoneyDivisionCondition;
-import lotto.validation.money.MoneyValidation;
-import lotto.validation.money.NumberCondition;
 import lotto.view.impl.LottoBuyView;
 
 public class Application {
@@ -19,11 +16,7 @@ public class Application {
     private static void run() {
         Money lottoPrice = Money.wons(1000);
         Store store = new Store(
-            new LottoBuyView(
-                    new MoneyValidation(
-                            new NumberCondition(), new MoneyDivisionCondition(lottoPrice)
-                    )
-            )
+            new LottoBuyView()
         );
 
         StatisticProducer statisticProducer = new StatisticProducer();
