+++ src/main/java/lotto/view/impl/LottoBuyView.java
@@ -3,7 +3,6 @@ package lotto.view.impl;
 import camp.nextstep.edu.missionutils.Console;
 import lotto.domain.Lottos;
 import lotto.domain.Money;
-import lotto.validation.Validation;
 import lotto.view.Inputable;
 import lotto.view.Outputable;
 import lotto.view.View;
@@ -16,20 +15,10 @@ public class LottoBuyView extends View implements Inputable, Outputable{
     private final static String BEGIN_MARKER = "[";
     private final static String END_MARKER = "]";
 
-    private Validation moneyValidation;
-
-    public LottoBuyView(Validation moneyValidation) {
-        this.moneyValidation = moneyValidation;
-    }
-
     @Override
     public Object inputView() {
         printView(moneyInputText);
-
-        String money = Console.readLine();
-        moneyValidation.validate(money);
-
-        return new Money(money);
+        return new Money(Console.readLine());
     }
 
     @Override
